Moved LinkedListAllocator from main.cc into separte files
