+++ BlockHeader.hh
@@ -1,0 +1,12 @@
+#pragma once
+#include <stddef.h>
+struct BlockHeader {
+    size_t size;
+    union {
+        struct
+        {
+            BlockHeader *prev, *next;
+        };
+        char data[1];
+    };
+};
