+++ LinkedListAllocator.cc
@@ -1,0 +1,215 @@
+
+
+#include "LinkedListAllocator.hh"
+
+#include <malloc.h>
+#include <stdio.h>
+
+#include <exception>
+#include <new>
+
+#include "BlockHeader.hh"
+#include "constants.hh"  // Почему автоформатирование делает такой странный порядок инклудов...................
+
+// Alias for Free header size
+static constexpr size_t F_hrader = sizeof(BlockHeader);
+// Alias for Allocated header size
+static constexpr size_t A_header = sizeof(size_t);
+
+char* buffer;
+BlockHeader* root;
+
+#if DEBUG
+// сумма байт всех запросов
+size_t bytes_allocated;
+// сумма занятых байт
+size_t bytes_used;
+size_t block_counter;
+size_t max_block_count;
+size_t block_size_distribution[LARGE_BUFFER_SIZE + 1] = {0};
+#endif
+
+LinkedListAllocator::LinkedListAllocator() {
+    buffer = static_cast<char*>(malloc(LARGE_BUFFER_SIZE));
+    // непрямое создание первого блока во весь буффер
+    // только в этом языке можно создать обьект без конструктора, лол
+    root->prev = root->next = root = reinterpret_cast<BlockHeader*>(buffer);
+    root->size = LARGE_BUFFER_SIZE;
+
+#if DEBUG
+    bytes_allocated = 0;
+    bytes_used = sizeof(BlockHeader);
+    block_counter = max_block_count = 1;
+    block_size_distribution[LARGE_BUFFER_SIZE] = 1;
+
+    printf(
+        "Linked List Allocator construction succesful\n\
+root: %p, buffer size: %d\n\n",
+        root, LARGE_BUFFER_SIZE);
+#endif
+}
+
+LinkedListAllocator::~LinkedListAllocator() {
+#if DEBUG
+    printf(
+        "Linked Aist Allocator decostructor log\n\
+root: addr: %p, size: %lu\n\
+bytes left allocated: %lu, bytes left used: %lu\n\
+sinle block left? %d, block count: %lu, max block count: %lu\n\
+block distribution:\n",
+        root, root->size,
+        bytes_allocated, bytes_used,
+        root->prev == root && root == root->next, block_counter, max_block_count);
+
+    for (size_t size = 0; size < LARGE_BUFFER_SIZE; ++size) {
+        size_t count = block_size_distribution[size];
+        if (count) {
+            printf("size %lu blocks %lu\n", size, count);
+        }
+    }
+#endif
+    free(buffer);
+}
+
+void LinkedListAllocator::remove_from_list(BlockHeader* rem) {
+    BlockHeader* prev = rem->prev;
+    BlockHeader* next = rem->next;
+    if (rem == root) {
+        if (rem == next) {
+            // список из 1 удоляемого элемента - забываем
+            root = nullptr;
+            return;
+        } else {
+            // потдержание доступа к списку
+            root = next;
+        }
+    }
+    prev->next = next;
+    next->prev = prev;
+    return;
+}
+
+char* LinkedListAllocator::allocate(size_t size) {
+    if (root == nullptr) {
+        printf("No Free blocks (or linked list access is lost)\n");
+        throw std::bad_alloc();
+    }
+
+    BlockHeader* cur = root;
+    do {
+        cur = cur->next;  // root - последний элемент итерации
+    } while (cur != root && cur->size < size + A_header);
+
+    if (cur->size < size + A_header) {
+        printf(
+            "No Block of sufficient size error\nBuffer size: %d\n\
+                last alloc request/last checked block capacity: %lu/%lu\n",
+            LARGE_BUFFER_SIZE, size, cur->size - A_header);
+        throw std::bad_alloc();
+    }
+
+    // от блока достаточного размера отрезаем новый блок требуемого размера
+    if (cur->size >= size + F_hrader + A_header) {
+        BlockHeader* cuted_block = reinterpret_cast<BlockHeader*>((char*)cur + cur->size - size - A_header);
+
+#if DEBUG
+        bytes_allocated += size;
+        bytes_used += size + A_header;
+        ++block_counter;
+        max_block_count = (block_counter > max_block_count) ? (block_counter) : (max_block_count);
+        --block_size_distribution[cur->size];
+        ++block_size_distribution[size + A_header];
+        ++block_size_distribution[cur->size - size - A_header];
+#endif
+
+        cur->size -= (cuted_block->size = size + A_header);
+        return cuted_block->data;
+    } else {
+        remove_from_list(cur);
+
+#if DEBUG
+        // не совсем правдо зато без утечек
+        bytes_allocated += cur->size - A_header;
+        bytes_used += cur->size - F_hrader;
+#endif
+
+        return cur->data;
+    }
+}
+
+// Я конечно сделал, но нужно допиливать. Безбожно медленно + по логированию очищается не всё.
+// (вообще не понял почему он не бросает segmentation segmentation falts).
+void LinkedListAllocator::deallocate(void* ptr) {
+    //!!! p - адресс на начала данных, не блока
+    BlockHeader* to_free = reinterpret_cast<BlockHeader*>((char*)ptr - A_header);
+#if DEBUG
+    bytes_allocated -= to_free->size - A_header;
+#endif
+    if (root == nullptr) {
+        root = to_free->prev = to_free->next = to_free;
+        return;
+    }
+
+    BlockHeader *cur = root, *prev_free = nullptr, *next_free = nullptr;
+    do {  // запоминаем граничащие свободные блоки
+        if ((char*)cur + cur->size == (char*)to_free) {
+            prev_free = cur;
+        }
+        if ((char*)to_free + to_free->size == (char*)cur) {
+            next_free = cur;
+        }
+    } while ((cur = cur->next) != root);
+
+    // далее обработка 4 случаев. (наличие/отсутствие)*(првого/левого) соседа для слияния
+    switch ((bool)prev_free << 1 | (bool)next_free) {
+    case 0:  // соседей нет => вставляем блок за корнем
+        (to_free->prev = root)->next = (to_free->next = root->next)->prev = to_free;
+#if DEBUG
+        bytes_used -= to_free->size - F_hrader;
+#endif
+        break;
+    case 1:  // только сосед справа => подменяем в списке соседа на себя и присоединяем его
+        (to_free->prev = next_free->prev)->next = (to_free->next = next_free->next)->prev = to_free;
+        to_free->size += next_free->size;
+
+#if DEBUG
+        bytes_used -= to_free->size;
+        --block_counter;
+        --block_size_distribution[to_free->size];
+        --block_size_distribution[next_free->size];
+        ++block_size_distribution[to_free->size + next_free->size];
+#endif
+
+        break;
+    case 2:  // только сосед слева => присоеденяемся к нему
+        prev_free->size += to_free->size;
+
+#if DEBUG
+        bytes_used -= to_free->size;
+        --block_counter;
+        --block_size_distribution[prev_free->size];
+        --block_size_distribution[to_free->size];
+        ++block_size_distribution[to_free->size + prev_free->size];
+#endif
+
+        break;
+    case 3:  // оба соседа => вырезаем правого и присоденяем себя и правого к левому
+        prev_free->size += to_free->size + next_free->size;
+        remove_from_list(next_free);
+
+#if DEBUG
+        bytes_used -= to_free->size + F_hrader;
+        block_counter -= 2;
+        --block_size_distribution[prev_free->size];
+        --block_size_distribution[to_free->size];
+        --block_size_distribution[next_free->size];
+        //++block_size_distribution[prev_free->size + to_free->size + next_free->size];
+#endif
+
+        break;
+    default:
+        printf("мы не знаем что это такое...\n");
+        throw std::bad_exception();
+        break;
+    }
+}
